Added finished visuals, finished Bernice Intro w/ camera movement
The dialogue system wasn't really compatible with cinemachine camera movement since it stayed in place, so I made subclasses to most of the dialogue system scripts for bernice-specific versions that would work on a canvas. The only changes made to the original scripts themselves are accessibility changes from private to protected for subclassing purposes.